+++ src/vmm/src/rpc_interface.rs
@@ -5,6 +5,8 @@ use std::fmt::{Display, Formatter};
 use std::result;
 use std::sync::{Arc, Mutex, MutexGuard};
 
+use guest_config::cpu::cpu_config::CustomCpuConfiguration;
+use guest_config::GuestConfigurationError;
 use logger::*;
 use mmds::data_store::{self, Mmds};
 use seccompiler::BpfThreadMap;
@@ -93,6 +95,8 @@ pub enum VmmAction {
     Pause,
     /// Repopulate the MMDS contents.
     PutMMDS(Value),
+    /// Configure the guest vCPU features.
+    PutCpuConfiguration(CustomCpuConfiguration),
     /// Resume the guest, by resuming the microVM VCPUs.
     Resume,
     /// Set the balloon device or update the one that already exists using the
@@ -134,6 +138,8 @@ pub enum VmmActionError {
     BootSource(BootSourceConfigError),
     /// The action `CreateSnapshot` failed.
     CreateSnapshot(CreateSnapshotError),
+    /// The action `ConfigureCpu` failed.
+    ConfigureCpu(GuestConfigurationError),
     /// One of the actions `InsertBlockDevice` or `UpdateBlockDevicePath`
     /// failed because of bad user input.
     DriveConfig(DriveError),
@@ -180,6 +186,7 @@ impl Display for VmmActionError {
             match self {
                 BalloonConfig(err) => err.to_string(),
                 BootSource(err) => err.to_string(),
+                ConfigureCpu(err) => err.to_string(),
                 CreateSnapshot(err) => err.to_string(),
                 DriveConfig(err) => err.to_string(),
                 InternalVmm(err) => format!("Internal Vmm error: {}", err),
@@ -427,6 +434,7 @@ impl<'a> PrebootApiController<'a> {
                 .load_snapshot(&config)
                 .map_err(VmmActionError::LoadSnapshot),
             PatchMMDS(value) => self.patch_mmds(value),
+            PutCpuConfiguration(cpu_config) => self.put_cpu_configuration(cpu_config),
             PutMMDS(value) => self.put_mmds(value),
             SetBalloonDevice(config) => self.set_balloon_device(config),
             SetVsockDevice(config) => self.set_vsock_device(config),
@@ -502,6 +510,13 @@ impl<'a> PrebootApiController<'a> {
             .update_vm_config(&cfg)
             .map(|()| VmmData::Empty)
             .map_err(VmmActionError::MachineConfig)
+    }
+
+    fn put_cpu_configuration(&mut self, cpu_config: CustomCpuConfiguration) -> ActionResult {
+        self.vm_resources
+            .configure_cpu(cpu_config)
+            .map(|()| VmmData::Empty)
+            .map_err(VmmActionError::ConfigureCpu)
     }
 
     fn set_vsock_device(&mut self, cfg: VsockDeviceConfig) -> ActionResult {
@@ -666,6 +681,7 @@ impl RuntimeApiController {
             | InsertBlockDevice(_)
             | InsertNetworkDevice(_)
             | LoadSnapshot(_)
+            | PutCpuConfiguration(_)
             | SetBalloonDevice(_)
             | SetVsockDevice(_)
             | SetMmdsConfiguration(_)
@@ -908,7 +924,7 @@ mod tests {
             self.vm_config.vcpu_count = machine_config.vcpu_count.unwrap();
             self.vm_config.mem_size_mib = machine_config.mem_size_mib.unwrap();
             self.vm_config.smt = machine_config.smt.unwrap();
-            self.vm_config.cpu_template = machine_config.cpu_template.unwrap();
+            self.vm_config.cpu_template = machine_config.cpu_template.as_ref().unwrap().clone();
             self.vm_config.track_dirty_pages = machine_config.track_dirty_pages.unwrap();
 
             Ok(())
@@ -994,6 +1010,17 @@ mod tests {
         pub fn locked_mmds_or_default(&mut self) -> MutexGuard<'_, Mmds> {
             let mmds = self.mmds_or_default();
             mmds.lock().expect("Poisoned lock")
+        }
+
+        pub fn configure_cpu(
+            &mut self,
+            cpu_config: CustomCpuConfiguration,
+        ) -> Result<(), GuestConfigurationError> {
+            // Update the CPU configuration on the VM
+            self.vm_config.cpu_template =
+                crate::vmm_config::machine_config::CpuFeaturesTemplate::CUSTOM(cpu_config);
+
+            Ok(())
         }
     }