+++ src/vmm/src/vstate/vcpu/x86_64.rs
@@ -258,14 +258,28 @@ impl KvmVcpu {
         vcpu_config: &VcpuConfig,
         mut cpuid: CpuId,
     ) -> std::result::Result<(), KvmVcpuConfigureError> {
+        // Set boot MSRs
+        let mut msr_boot_entries = arch::x86_64::msr::create_boot_msr_entries();
+
         // If a template is specified, get the CPUID template, else use `cpuid`.
-        let mut config_cpuid = match vcpu_config.cpu_template {
+        let mut config_cpuid = match &vcpu_config.cpu_template {
             #[cfg(feature = "t2")]
             CpuFeaturesTemplate::T2 => cpuid::Cpuid::T2,
             #[cfg(feature = "t2s")]
             CpuFeaturesTemplate::T2S => cpuid::Cpuid::T2S,
             #[cfg(feature = "c3")]
             CpuFeaturesTemplate::C3 => cpuid::Cpuid::C3,
+            CpuFeaturesTemplate::CUSTOM(cpu_config) => {
+                // Build holistic configuration
+                // Currently includes:
+                // * Base CPUID configuration
+                // * CPUID overrides
+                // * MSR overrides
+                guest_config::cpu::cpu_symbolic_engine::configure_cpu_features(
+                    &mut msr_boot_entries,
+                    cpu_config,
+                )
+            }
             // If a template is not supplied we use the given `cpuid` as the base.
             CpuFeaturesTemplate::None => {
                 cpuid::Cpuid::try_from(cpuid::RawCpuid::from(cpuid.clone()))?
@@ -308,8 +322,6 @@ impl KvmVcpu {
             .set_cpuid2(&cpuid)
             .map_err(KvmVcpuConfigureError::SetCpuid)?;
 
-        // Set MSRs
-        let mut msr_boot_entries = arch::x86_64::msr::create_boot_msr_entries();
         #[cfg(feature = "t2s")]
         if vcpu_config.cpu_template == CpuFeaturesTemplate::T2S {
             for msr in cpuid::t2s::msr_entries_to_save() {